Update chartjs.py
change data = serializers.ListField(child=serializers.IntegerField())
to data = serializers.ListField(child=serializers.FloatField())

+++ djgentelella/chartjs.py
@@ -6,7 +6,7 @@ class DataSetSerializer(serializers.Serializer):
     label = serializers.CharField()
     backgroundColor = serializers.CharField(required=False)
     borderColor = serializers.CharField(required=False)
-    data = serializers.ListField(child=serializers.IntegerField())
+    data = serializers.ListField(child=serializers.FloatField())
     fill = serializers.BooleanField(default=False, required=False)
     borderWidth = serializers.IntegerField(default=0, required=False)
     stack = serializers.CharField(required=False)
